chore: Organize tests
