+++ py-polars/tests/unit/operations/col_selection/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for column selection operations."""
