+++ py-polars/tests/unit/operations/join/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for join operations."""
