+++ py-polars/tests/unit/operations/join/test_cross_join.py
