+++ py-polars/tests/unit/operations/predicates/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for Polars boolean predicate expressions."""
