+++ py-polars/tests/unit/operations/reshape/test_explode.py
